requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=15.0
nuscenes-devkit>=1.1
scikit-learn>=1.4
umap-learn>=0.5
matplotlib>=3.8
shapely>=2.0
